pyproject.toml
[project]
name = "fastapi_fabric_cosmos_db_rag"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sumsam Ali", email = "sumsamali.cs@gmail.com"},
]
dependencies = ["fastapi>=0.124.0", "uvicorn>=0.38.0", "openai>=2.9.0", "pydantic>=2.12.5", "pydantic-settings>=2.12.0", "azure-cosmos>=4.14.2", "python-dotenv>=1.2.1", "pytest>=9.0.2", "pytest-asyncio>=1.3.0", "httpx>=0.28.1", "aiohttp>=3.13.2", "azure-identity>=1.25.1"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"